refactoring pr
