+++ package.json
@@ -27,11 +27,10 @@
     "@babel/plugin-proposal-class-properties": "^7.8.3",
     "@babel/plugin-proposal-object-rest-spread": "^7.9.6",
     "@babel/preset-env": "^7.9.6",
-    "anyproxy": "^4.1.2",
     "chai": "^4.2.0",
     "chai-subset": "^1.6.0",
     "mocha": "^7.1.2",
-    "nodemon": "^2.0.3",
+    "nodemon": "^2.0.4",
     "pkg": "^4.4.8",
     "read-last-lines": "^1.7.2"
   },
