+++ src/connectors/connectorRIS.js
@@ -45,10 +45,7 @@ export default class ConnectorRIS extends Connector{
         this.pingInterval = 5000;
         this._defaultReconnectTimeout = 10000;
         this.reconnectTimeout = this._defaultReconnectTimeout;
-        this.agent = null;
-        if (!this.params.noProxy) {
-            this.agent = env.agent;
-        }
+        this.agent = env.agent;
 
         setInterval(this._ping, this.pingInterval);
 
@@ -91,10 +88,13 @@ export default class ConnectorRIS extends Connector{
     connect = () =>
         new Promise((resolve, reject) => {
             try {
-                this.ws = new WebSocket(this.url, {
-                    agent: this.agent,
+                const wsOptions = {
                     perMessageDeflate: this.params.perMessageDeflate
-                });
+                };
+                if (!this.params.noProxy && this.agent) {
+                    wsOptions.agent = this.agent;
+                }
+                this.ws = new WebSocket(this.url, wsOptions);
 
                 this.ws.on('message', this._messageToJson);
                 this.ws.on('close', (error) => {
