+++ src/processMonitors/uptime.js
@@ -30,11 +30,16 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
+import axios from "axios";
+import env from "../env";
+
 export default class Uptime {
 
     constructor(connectors, params){
         this.connectors = connectors;
         this.params = params;
+        axios.defaults.agent = env.agent;
+        this.axios = axios;
     };
