+++ src/reports/report.js
@@ -33,6 +33,7 @@
 
 import moment from "moment";
 import brembo from "brembo";
+import axios from "axios";
 
 export default class Report {
 
@@ -48,6 +49,11 @@ export default class Report {
                 return this.report(message, content);
             });
         }
+
+        if (!this.params.noProxy && env.agent) {
+            axios.defaults.agent = env.agent;
+        }
+        this.axios = axios;
     }
 
     getBGPlayLink = (prefix, start, end, instant = null, rrcs = [0,1,2,5,6,7,10,11,13,14,15,16,18,20]) => {
