+++ src/reports/reportSlack.js
@@ -36,9 +36,6 @@ export default class ReportSlack extends ReportHTTP {
 
     constructor(channels, params, env) {
         const templates = {};
-        if (!this.params.noProxy) {
-            axios.defaults.agent = env.agent;
-        }
 
         const getTemplateItem = (color) => {
            return JSON.stringify({
