+++ src/reports/reportWebex.js
@@ -31,7 +31,6 @@
  */
 
 import Report from "./report";
-import axios from "axios";
 
 export default class ReportWebex extends Report {
 
@@ -54,14 +53,11 @@ export default class ReportWebex extends Report {
                 });
             }
         }
-        if (!this.params.noProxy) {
-            axios.defaults.agent = env.agent;
-        }
     }
 
     _sendWebexMessage = (url, message, content) => {
 
-        axios({
+        this.axios({
             url: url,
             method: "POST",
             resposnseType: "json",
